Finalize the react dashboard app
