+++ .gitignore
@@ -2,3 +2,7 @@
 */target/*
 .DS_Store
 *.iml
+
+dist/
+node_modules/
+tmp/
