+++ dashboard/react-app/src/App.js
@@ -1,24 +1,6 @@
-/*
- * Copyright (c) 2019, Entgra (pvt) Ltd. (http://entgra.io) All Rights Reserved.
- *
- * Entgra (pvt) Ltd. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
 import React from "react";
 import "antd/dist/antd.less";
-import RouteWithSubRoutes from "./components/RouteWithSubRoutes";
+import RouteWithSubRoutes from "./js/RouteWithSubRoutes";
 import {
     BrowserRouter as Router,
     Redirect, Switch,
