+++ dashboard/react-app/src/pages/Dashboard/Home/Home.js
@@ -1,0 +1,15 @@
+import React from "react";
+
+class Home extends React.Component{
+    render() {
+        return(
+            <div>
+                <div style={{background: '#f0f2f5', padding: 24, minHeight: 780}}>
+                    <h1>Home</h1>
+                </div>
+            </div>
+        )
+    }
+}
+
+export default Home;
