--- dashboard/react-app/src/pages/dashboard/Home.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-class Home extends React.Component{
-    render() {
-        return(
-            <h1>Home</h1>
-        )
-    }
-}
-
-export default Home;
